backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Project Management MVP"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "httpx",
    "pytest",
]

[tool.pytest.ini_options]
pythonpath = ["."]